Gather embedded files + write embed module in separate steps
Breaking this block means that multiple invocations of the `--embed`
flag will no longer cause the `embed.go` file to be written once per
invocation (the output of which will be identical).

Note that I also reordered the statements for each iteration of
gathering the embedded files. Before, `copy()` was called before the
`Stat()` check, which likely meant that a nonexistent Dir would error
at the `copy()` call before the more-descriptive error from the `Stat()`
check could be returned.

Reordering these checks and associated logs *could* be a breaking change
if any consumers are checking log output in some kind of automated test.